refactor: create entity_extractor module
