+++ aether-kernel/aether/kernel/api/urls.py
@@ -37,5 +37,5 @@ router.register('entities', views.EntityViewSet)
 
 
 urlpatterns = router.urls + [
-    url(r'^validate-mappings/$', views.validate_mappings, name='validate-mappings'),
+    url(r'^validate-mappings/$', views.validate_mappings_view, name='validate-mappings'),
 ]
