+++ aether-kernel/aether/kernel/api/validators.py
@@ -16,13 +16,19 @@
 # specific language governing permissions and limitations
 # under the License.
 
+import collections
 import json
+import uuid
 
 from django.core.exceptions import ValidationError
 from django.utils.translation import ugettext as _
 
 from jsonschema.validators import Draft4Validator
+
+from spavro.io import validate
 from spavro.schema import parse, SchemaParseException
+
+from . import avro_tools
 
 
 MESSAGE_REQUIRED_ID = _('A schema is required to have a field "id" of type "string"')
@@ -66,6 +72,11 @@ MAPPING_DEFINITION_SCHEMA = {
 }
 
 mapping_definition_validator = Draft4Validator(MAPPING_DEFINITION_SCHEMA)
+
+EntityValidationResult = collections.namedtuple(
+    'EntityValidationResult',
+    ['validation_errors', 'entities'],
+)
 
 
 def validate_avro_schema(value):
@@ -138,3 +149,64 @@ def validate_schemas(value):
         validate_schema_definition(schema)
 
     return value
+
+
+def validate_entity_payload(schema_definition, payload):
+    # Use spavro to validate payload against the linked schema
+    try:
+        avro_schema = parse(json.dumps(schema_definition))
+        valid = validate(avro_schema, payload)
+        if not valid:
+            msg = _('Extracted record did not conform to registered schema')
+            raise ValidationError(msg)
+        return True
+    except Exception as err:
+        raise ValidationError(str(err))
+
+
+def validate_entity_payload_id(entity_payload):
+    id_ = entity_payload.get('id', None)
+    try:
+        uuid.UUID(id_, version=4)
+        return None
+    except (ValueError, AttributeError, TypeError):
+        return {'description': _('Entity id "{}" is not a valid uuid').format(id_)}
+
+
+def validate_avro(schema, datum):
+    result = avro_tools.AvroValidator(
+        schema=parse(json.dumps(schema)),
+        datum=datum,
+    )
+    errors = []
+    for error in result.errors:
+        errors.append({
+            'description': avro_tools.format_validation_error(error),
+        })
+    return errors
+
+
+def validate_entities(entities, schemas):
+    validation_errors = []
+    validated_entities = collections.defaultdict(list)
+    for entity_name, entity_payloads in entities.items():
+        for entity_payload in entity_payloads:
+            entity_errors = []
+            id_error = validate_entity_payload_id(entity_payload)
+            if id_error:
+                entity_errors.append(id_error)
+            schema_definition = schemas[entity_name]
+            avro_validation_errors = validate_avro(
+                schema=schema_definition,
+                datum=entity_payload,
+            )
+            entity_errors.extend(avro_validation_errors)
+            if entity_errors:
+                validation_errors.extend(entity_errors)
+            else:
+                validated_entities[entity_name].append(entity_payload)
+
+    return EntityValidationResult(
+        validation_errors=validation_errors,
+        entities=validated_entities,
+    )
